requirements.txt
ipython==8.14.0
matplotlib==3.6.3
numpy==1.24.1
pyserial==3.5
scikit_learn==1.3.2
scipy==1.11.4
